renaming to app_logging
