+++ api/blueprints/drbFulfill.py
@@ -7,7 +7,7 @@ from ..db import DBClient
 from ..utils import APIUtils
 from ..validation_utils import is_valid_numeric_id
 from managers import S3Manager
-from logging import logger
+from app_logging import logger
 
 fulfill = Blueprint('fulfill', __name__, url_prefix='/fulfill')
 response_type = 'fulfill'
