+++ managers/coverFetchers/googleBooksFetcher.py
@@ -2,7 +2,7 @@ import os
 import requests
 from requests.exceptions import ReadTimeout, HTTPError
 
-from logging import logger
+from app_logging import logger
 from managers.coverFetchers.abstractFetcher import AbstractFetcher
