+++ mappings/marc.py
@@ -1,7 +1,7 @@
 import re
 
 from .core import Core
-from logging import logger
+from app_logging import logger
 
 
 class MARCMapping(Core):
