+++ processes/dbMaintenance.py
@@ -1,5 +1,5 @@
 from .core import CoreProcess
-from logging import logger
+from app_logging import logger
 
 
 class DatabaseMaintenanceProcess(CoreProcess):
